# Various fixes related to styles handling on load